% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-activities.R
\name{activities}
\alias{activities}
\title{Append activity data to a discoveryengine definition}
\usage{
activities(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\description{
Append activity data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_activities = activities(wealthy)
display(wealthy_activities)

}
