% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-all.R
\name{all_chunks}
\alias{all_chunks}
\title{Append all available chunks to a discoveryengine definition}
\usage{
all_chunks(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with columns from all available chunks
See \code{\link{basic_bio}},
\code{\link{degrees}}
\code{\link{employment}}
\code{\link{rating}}
\code{\link{screening}}
\code{\link{activities}}
\code{\link{interests}}
\code{\link{prospect}}
\code{\link{giving}}
}
\description{
Append all available chunks to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_data = all_chunks(wealthy)
display(wealthy_data)

}
