% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-geocode.R, R/disco_append.R
\name{geocode}
\alias{geocode}
\alias{append}
\alias{basic_bio}
\alias{capacity}
\alias{activities}
\alias{giving}
\alias{degrees}
\alias{employment}
\alias{prospect}
\alias{append}
\title{Append data to a discoveryengine definition}
\usage{
geocode(constituency)

basic_bio(constituency)

capacity(constituency)

activities(constituency)

giving(constituency)

degrees(constituency)

employment(constituency)

prospect(constituency)

append(constituency, ...)
}
\arguments{
\item{constituency}{A discoveryengine definition}

\item{...}{One or more sections to be appended}
}
\description{
Append data to a discoveryengine definition
}
\examples{
library(discoveryengine)
wealthy = has_capacity(1)
display(basic_bio(wealthy))

display(append(wealthy, basic_bio, capacity, activities))

}
