% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-capacity.R
\name{capacity}
\alias{capacity}
\title{Append capacity data to a discoveryengine definition}
\usage{
capacity(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\description{
Append capacity data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_capacity = capacity(wealthy)
display(wealthy_capacity)

}
