% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contacts_by_unit.R
\name{contacts_by_unit}
\alias{contacts_by_unit}
\title{Append last three contacts by unit to a discoveryengine definition}
\usage{
contacts_by_unit(constituency, unit)
}
\arguments{
\item{constituency}{A discoveryengine definition}

\item{unit}{Unit code}
}
\value{
data frame with the following columns: entity id, last contact date, last contact, second last contact date, second last contact, third last contact date, third last contact
}
\description{
Append last three contacts by unit to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
cal_performances_contacts = contacts_by_unit(wealthy, "CP")
display(cal_performances_contacts)

}
