% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-giving.R
\name{custom_ca_giving}
\alias{custom_ca_giving}
\alias{ca_giving_to_proposition}
\alias{ca_giving_to_candidate}
\title{Detailed CA campaign giving summaries in custom chunks}
\usage{
ca_giving_to_proposition(..., from = NULL, to = NULL)

ca_giving_to_candidate(..., from = NULL, to = NULL)
}
\arguments{
\item{...}{One or more codes (candidate/proposition)}

\item{from}{Start date, in the format YYYYMMDD}

\item{to}{End date, in the format YYYYMMDD}
}
\description{
These functions are used inside the \code{\link{custom}} function, to create
detailed summaries of California state level political contributions. They have a similar
interface to the ca_gave_to_* widgets in the Discovery Engine.
}
\examples{
wealthy = has_capacity(1)

## summary of all giving to CA state level campaigns to candidates
wealthy \%>\%
  custom(
    ca_giving = ca_giving_to_candidate()
  )

## find donors to Kamala Harris (whose code is CA770144) and append amounts
ca_gave_to_candidate(CA770144) \%>\%
custom(
harris_giving = ca_giving_to_candidate(CA770144)
)

}
\seealso{
\code{\link{custom}}
}
