% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fec-giving.R
\name{custom_fec_giving}
\alias{custom_fec_giving}
\alias{fec_giving_to_committee}
\alias{fec_giving_to_category}
\alias{fec_giving_to_candidate}
\title{Detailed FEC summaries in custom chunks}
\usage{
fec_giving_to_committee(..., from = NULL, to = NULL)

fec_giving_to_category(..., from = NULL, to = NULL)

fec_giving_to_candidate(..., from = NULL, to = NULL)
}
\arguments{
\item{...}{One or more codes (committee/candidate/category) or synonyms}

\item{from}{Start date, in the format YYYYMMDD}

\item{to}{End date, in the format YYYYMMDD}
}
\description{
These functions are used inside the \code{\link{custom}} function, to create
detailed summaries of federal political contributions. They have a similar
interface to the fec_gave_to_* widgets in the Discovery Engine.
}
\examples{
wealthy = has_capacity(1)

## summary of all giving to FEC
wealthy \%>\%
  custom(
    fec_giving = fec_giving_to_committee()
  )

## giving to environmental PACS
wealthy \%>\%
  custom(
    env_giving = fec_giving_to_category(environmental_policy)
  )

}
\seealso{
\code{\link{custom}}
}
