% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-giving.R
\name{custom_giving}
\alias{custom_giving}
\alias{giving_to_area}
\alias{giving_to_department}
\alias{giving_to_fund_type}
\alias{giving_to_fund}
\title{Detailed giving summaries in custom chunks}
\usage{
giving_to_area(..., from = NULL, to = NULL)

giving_to_department(..., from = NULL, to = NULL)

giving_to_fund_type(..., from = NULL, to = NULL)

giving_to_fund(..., from = NULL, to = NULL)
}
\arguments{
\item{...}{One or more codes (area/department/type, etc) or synonyms}

\item{from}{Start date, in the format YYYYMMDD}

\item{to}{End date, in the format YYYYMMDD}
}
\description{
These functions are used inside the \code{\link{custom}} function, to create
detailed giving summaries. They have a similar interface to the giving widgets
in the Discovery Engine.
}
\examples{
wealthy = has_capacity(1)

## get total giving to campus in FY15 and FY16
## as with the disco engine, i can get giving anywhere on campus
## by not specifying an area
wealthy \%>\%
  custom(
    fy15_giving = giving_to_area(from = 20140701, to = 20150630),
    fy16_giving = giving_to_area(from = 20150701, to = 20160630)
  )

## total giving to neuro/brain related funds:
neuro_fund = fund_text_contains("neuro*", "brain")
wealthy \%>\%
  custom(neuro_giving = giving_to_fund(neuro_fund))

}
\seealso{
\code{\link{custom}}
}
