% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-degrees.R
\name{degrees}
\alias{degrees}
\title{Append degree data to a discoveryengine definition}
\usage{
degrees(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, undergrad degree, graduate degree, degrees, preferred school code, preferred school, class campaign year, spouse undergraduate degree, spouse graduate degree, spouse degrees, spouse class campaign year
}
\description{
Append degree data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_degrees = degrees(wealthy)
display(wealthy_degrees)

}
