% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-employment.R
\name{employment}
\alias{employment}
\title{Append employment data to a discoveryengine definition}
\usage{
employment(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\description{
Append employment data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_employment = employment(wealthy)
display(wealthy_employment)

}
