% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-flag.R
\name{flag}
\alias{flag}
\title{Flag a constitutency}
\usage{
flag(constituency)
}
\arguments{
\item{constituency}{A constituency defined with the Disco Engine}
}
\description{
Use within a \code{\link{custom}} chunk to "flag" a constituency with a
Y/N column. See examples
}
\examples{
## among those with a rating of 1, who attended an event in 2016?
wealthy = has_capacity(1)
event_attendee = attended_event(from = 20160101, to = 2016)

wealthy \%>\%
  custom(
    event_attendee_16 = flag(event_attendee)
  )

}
\seealso{
\code{\link{custom}}
}
