% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-giving.R
\name{giving}
\alias{giving}
\title{Append giving data to a discoveryengine definition}
\usage{
giving(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\description{
Append giving data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_giving = giving(wealthy)
display(wealthy_giving)

}
