% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-interests.R
\name{interests}
\alias{interests}
\title{Append interest to a discoveryengine definition}
\usage{
interests(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, affiliations, interests, philanthropic interests, philanthropic affinities, philanthropic organizations, committees
}
\description{
Append interest to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_interests = interests(wealthy)
display(wealthy_interests)

}
