% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-prospect.R
\name{prospect}
\alias{prospect}
\title{Append prospect data to a discoveryengine definition}
\usage{
prospect(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\description{
Append prospect data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_prospect = prospect(wealthy)
display(wealthy_prospect)

}
