% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-robo-rating.R
\name{robo_rating}
\alias{robo_rating}
\title{Append automated capacity rating to a discoveryengine definition}
\usage{
robo_rating(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, median_income, fec_matched_giving, hh_fec_matched_giving, ca_matched_giving, hh_ca_matched_giving, has_sec, is_director, is_officer, is_ten_percenter, hh_has_sec, hh_is_director, hh_is_officer, hh_is_ten_percenter, sec_link, hh_sec_links
}
\description{
Append automated capacity rating to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy \%>\% screening \%>\% display
}
