% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-show-chunks.R
\name{show_chunks}
\alias{show_chunks}
\title{Browse the available chunk names and outputs}
\usage{
show_chunks()
}
\description{
Pops up an interactive table that allows you to sort and search through a
listing of all available chunks with outputs for each
}
