% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-all.R
\name{all_chunks}
\alias{all_chunks}
\title{Append all available chunks to a discoveryengine definition}
\usage{
all_chunks(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with columns from all available chunks
See \code{\link[discoappend]{basic_bio}},
\code{\link[discoappend]{degrees}},
\code{\link[discoappend]{employment}},
\code{\link[discoappend]{rating}},
\code{\link[discoappend]{screening}},
\code{\link[discoappend]{activities}},
\code{\link[discoappend]{interests}},
\code{\link[discoappend]{prospect}},
\code{\link[discoappend]{giving}}
}
\description{
Append all available chunks to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_data = all_chunks(wealthy)
display(wealthy_data)

}
