% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-geocode.R, R/append.R
\name{geocode}
\alias{geocode}
\alias{append}
\title{Append data to a discoveryengine definition}
\usage{
geocode(constituency)

append(constituency, ...)
}
\arguments{
\item{constituency}{A discoveryengine definition}

\item{...}{One or more chunks to be appended}
}
\description{
Append data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_data = append(wealthy, basic_bio, capacity, giving)
display(wealthy_data)

}
