% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-basic-bio.R
\name{basic_bio}
\alias{basic_bio}
\title{Append basic bio data to a discoveryengine definition}
\usage{
basic_bio(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\description{
Append basic bio data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_bio = basic_bio(wealthy)
display(wealthy_bio)

}
