% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-degrees.R
\name{degrees}
\alias{degrees}
\title{Append degree data to a discoveryengine definition}
\usage{
degrees(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\description{
Append degree data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_degrees = degrees(wealthy)
display(wealthy_degrees)

}
