% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-giving.R
\name{giving}
\alias{giving}
\title{Append giving data to a discoveryengine definition}
\usage{
giving(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, lifetime giving, largest gift, largest gift date, largest gift area, last gift, last gift date, last gift area, average gift, outstanding pledges
}
\description{
Append giving data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_giving = giving(wealthy)
display(wealthy_giving)

}
