% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-prospect.R
\name{prospect}
\alias{prospect}
\title{Append prospect data to a discoveryengine definition}
\usage{
prospect(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, active major proposal flg, active annual proposal flg, proposals, primary manager, primary manager office, last 3 events, event count, last contact, university signature
}
\description{
Append prospect data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_prospect = prospect(wealthy)
display(wealthy_prospect)

}
