% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-capacity.R
\name{rating}
\alias{rating}
\title{Append rating data to a discoveryengine definition}
\usage{
rating(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, capacity rating code, capacity rating description, inclination rating description, builder of berkeley, implied capacity score, implied capacity description, major gift score, major gift description
}
\description{
Append rating data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_ratings = rating(wealthy)
display(wealthy_ratings)

}
