% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-capacity.R
\name{rating}
\alias{rating}
\title{Append rating data to a discoveryengine definition}
\usage{
rating(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, capacity rating, capacity rating date, inclination rating, inclination rating date, implied capacity score, implied capacity description, implied capacity score date, major gift score, major gift description, major gift score date, CNR score, CNR score description, CNR score date, Gift Planning score, Gift Planning score description, Gift Planning score date, Engineering score, Engineering score description, Engineering score date, Haas score, Haas score description, Haas score date
}
\description{
Append rating data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy_ratings = rating(wealthy)
display(wealthy_ratings)

}
