% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk-screening.R
\name{screening}
\alias{screening}
\title{Append external screening data to a discoveryengine definition}
\usage{
screening(constituency)
}
\arguments{
\item{constituency}{A discoveryengine definition}
}
\value{
data frame with the following columns: entity id, acs_median_income, fec_matched_giving, hh_fec_matched_giving
}
\description{
Append external screening data to a discoveryengine definition
}
\examples{
wealthy = has_capacity(1)
wealthy \%>\% screening \%>\% display
}
